+++ docs/_pages/changelog.markdown
@@ -6,6 +6,7 @@ permalink: changelog/
 
 This page tracks major changes to the site, in reverse chronological order:
 
+- Added calc_overlaps.py script (08-21-23)
 - Edited the rest of the site (08-21-23)) 
 - Editing the home page (08-20-23)
 - Added CO (08-20-23)